separate plantspecs components
